% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abilityData.R
\docType{data}
\name{abilityData}
\alias{abilityData}
\title{Data on Pokemon Abilities}
\format{
dataframe
}
\usage{
abilityData
}
\description{
Contains data on Pokemon id numbers and names
}
\keyword{datasets}
