% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate_Max_Stats.R
\name{est_max}
\alias{est_max}
\title{Estimate the maximum possible values of any Pokemon.}
\usage{
est_max(name_list, reduce = FALSE, level = 50)
}
\arguments{
\item{name_list}{List of Pokemon names.}

\item{reduce}{Logical value to state whether to reduce the stats shown.}
}
\value{
A dataframe of estimate battle stats.
}
\description{
Estimate the maximum possible values of any Pokemon.
}
