% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate_Min_Stats.R
\name{est_min}
\alias{est_min}
\title{Estimate the minimum possible values of any Pokemon.}
\usage{
est_min(name_list, reduce = FALSE, level = 50)
}
\arguments{
\item{name_list}{List of Pokemon names.}

\item{reduce}{Logical value to state whether to reduce the stats shown.}
}
\value{
A dataframe of estimate battle stats.
}
\description{
Estimate the minimum possible values of any Pokemon.
}
