% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate_Value_Stat.R
\name{est_val}
\alias{est_val}
\title{Estimate single stat using specified values and nature (values will be applied to all 6 stats)}
\usage{
est_val(name_list, IV = 15, EV = 100, level = 50, Nature = 1)
}
\arguments{
\item{name_list}{}

\item{IV}{}

\item{EV}{}

\item{level}{}

\item{Nature}{}
}
\description{
Estimate single stat using specified values and nature (values will be applied to all 6 stats)
}
