% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchangeTeamSlot.R
\name{exchangeTeamSlot}
\alias{exchangeTeamSlot}
\title{exchange team slot}
\usage{
exchangeTeamSlot(
  team,
  replace = 1,
  newPokemon = data.frame(matrix(nrow = 33, ncol = 1))
)
}
\arguments{
\item{team}{Data Frame, the team that has a pokemon the user would like to have removed or replaced}

\item{replace}{integer, the team slot to be altered. Default is slot 1}

\item{name}{Data frame, the pokemon that should be placed in slot 'replace' on the team. Default is blank}
}
\value{
team
}
\description{
Removes or replaces a pokemon on a pokemon team data frame
}
\examples{
\dontrun{
exchangeTeamSlot(team,replace,newPokemon)
}
}
\keyword{team,remove,replace}
