% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePokemon.R
\name{generatePokemon}
\alias{generatePokemon}
\title{generate pokemon data frame}
\usage{
generatePokemon(
  targetPokemon,
  targetIVs = c(31, 31, 31, 31, 31, 31),
  targetEVs = c(0, 0, 0, 0, 0, 0),
  targetNature = "serious",
  targetMoveset = c(1, 1, 1, 1),
  targetAbility = 1,
  targetLvl = 50,
  targetTera = ""
)
}
\arguments{
\item{targetPokemon}{string - the identifier string of the pokemon to generate}

\item{targetIVs}{List of integers - the desired set of individual values for the target pokemon. Default is all IVs set to 31}

\item{targetNature}{string - the desired nature for the target pokemon. Default nature is Serious}

\item{targetMoveset}{List of integers - the desired moveset for the target pokemon. Moves should be in listed by their numerical identifiers. Default is all moves set to 1}

\item{targetAbility}{Integer - the desired ability for the target pokemon. This should be in the format 1, 2, or 3. Default is 1}

\item{targetLvl}{Integer - the desired level for the target pokemon. Should be in the range [1,100]. Default is 50}

\item{targetTera}{String - the desired terastallization type for the target pokemon. Default value is "" and returns a random type from among the target pokemon's types}

\item{targetEvs}{list of integers - the desired set of effort values for the target pokemon. Defulat is all EVs set to 0}
}
\value{
pokemonDataFrame
}
\description{
Generates a pokemon data frame based on the arguments supplied by the user
}
\examples{
\dontrun{
generatePokemon(targetPokemon,targetIVs=c(31,31,31,31,31,31),targetEVs=c(0,0,0,0,0,0),targetNature="serious",targetMoveset=c(1,1,1,1),targetAbility=1,targetLvl=50,targetTera="")
}
}
\keyword{new,generate,pokemon}
