% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpCalc.R
\name{hpCalc}
\alias{hpCalc}
\title{hit points calculator}
\usage{
hpCalc(poke, hpIV = 31, hpEV = 0, lvl = 50)
}
\arguments{
\item{poke}{string - the identifier for the pokemon upon the the calculation will be performed}

\item{hpIV}{integer - the hp individual value for the pokemon in the range [0,31]. Default value is 31}

\item{hpEV}{integer - the hp effort value for the pokemon, in the range [0,252]. Default value is 0}

\item{lvl}{integer - the level of the pokemon, in the range [1,100]. Default value is 50}
}
\value{
description
}
\description{
Calculates the hit point (HP) stat for a given pokemon
}
\examples{
\dontrun{
hpCalc(poke,hpIV,hpEV,lvl)
}
}
\keyword{hp,hit}
\keyword{points,calculate,stats}
