% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natureData.R
\docType{data}
\name{natureData}
\alias{natureData}
\title{Data on Pokemon Natures}
\format{
dataframe
}
\usage{
natureData
}
\description{
Contains data on all 25 Pokemon natures, including increased and decreased stats, flavor preference, nature name, and nature number
}
\keyword{datasets}
