% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newTeam.R
\name{newTeam}
\alias{newTeam}
\title{new pokemon team}
\usage{
newTeam(slot1 = "", slot2 = "", slot3 = "", slot4 = "", slot5 = "", slot6 = "")
}
\arguments{
\item{slot1}{string - the pokemon identifier desired for the team slot 1 pokemon. Random pokemon will be chosen if left blank.}

\item{slot2}{string - the pokemon identifier desired for the team slot 2 pokemon. Random pokemon will be chosen if left blank.}

\item{slot3}{string - the pokemon identifier desired for the team slot 3 pokemon. Random pokemon will be chosen if left blank.}

\item{slot4}{string - the pokemon identifier desired for the team slot 4 pokemon. Random pokemon will be chosen if left blank.}

\item{slot5}{string - the pokemon identifier desired for the team slot 5 pokemon. Random pokemon will be chosen if left blank.}

\item{slot6}{string - the pokemon identifier desired for the team slot 6 pokemon. Random pokemon will be chosen if left blank.}
}
\value{
teamFrame
}
\description{
Creates a new pokemon team data frame
}
\examples{
\dontrun{
newTeam(slot1,slot2,slot3,slot4,slot5,slot6)
}
}
\keyword{create,team,new}
