% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_Speed_Outcome.R
\name{outsp}
\alias{outsp}
\title{Predict Speed Outcomes}
\usage{
outsp(attacker, defender, attacker_level = 50, defender_level = 50)
}
\arguments{
\item{attacker}{}

\item{defender}{}
}
\description{
Minimum speed means 0 EV and neutral nature. Maximum Speed means 252 EV and beneficial nature.
}
