% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokedex.R
\docType{data}
\name{pokedex}
\alias{pokedex}
\title{data(pokedex)}
\format{
An object of class \code{data.frame} with 800 rows and 13 columns.
}
\usage{
pokedex
}
\description{
data(pokedex)
}
\keyword{datasets}
