% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemonData.R
\docType{data}
\name{pokemonData}
\alias{pokemonData}
\title{Data on Pokemon attributes}
\format{
dataframe
}
\usage{
pokemonData
}
\description{
Contains information on all pokemon stats including base stats, types, abilities, height, and weight
}
\keyword{datasets}
