% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemoves.R
\docType{data}
\name{pokemoves}
\alias{pokemoves}
\title{data(pokemoves)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 915 rows and 8 columns.
}
\usage{
pokemoves
}
\description{
data(pokemoves)
}
\keyword{datasets}
