% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_Damage_Outcomes.R
\name{predic}
\alias{predic}
\title{Predict Damage Outcomes}
\usage{
predic(
  attacker,
  move,
  defender,
  attacker_level = 50,
  defender_level = 50,
  q = TRUE
)
}
\arguments{
\item{attacker}{}

\item{move}{}

\item{defender}{}

\item{level}{}

\item{qpredic}{}
}
\description{
Predict damage outcomes of defenders with mid defensive and HP EVs. Assumes the attacker has max attacking EVs and a beneficial nature.
}
