% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict_Damage_Outcomes_Randoms.R
\name{predic2}
\alias{predic2}
\title{Predict Turn Outcomes for Randoms}
\usage{
predic2(
  attacker,
  move,
  defender,
  attacker_level = 50,
  defender_level = 50,
  q = TRUE
)
}
\arguments{
\item{attacker}{}

\item{move}{}

\item{defender}{}

\item{level}{}

\item{qpredic}{}
}
\description{
Predict turn outcomes assuming all pokemon have 85 EV spread, 31 ICs, and neutral natures.
}
