% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regD.R
\docType{data}
\name{regD}
\alias{regD}
\title{Regulation D Eligible Pokemon}
\format{
dataframe
}
\usage{
regD
}
\description{
A data set containing the national dex numbers and identifiers for all regulation d eligible pokemon
}
\keyword{datasets}
