% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistances.R
\docType{data}
\name{resistances}
\alias{resistances}
\title{data(resistances)}
\format{
An object of class \code{data.frame} with 801 rows and 41 columns.
}
\usage{
resistances
}
\description{
data(resistances)
}
\keyword{datasets}
