% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statCalc.R
\name{statCalc}
\alias{statCalc}
\title{pokemon stat calculator}
\usage{
statCalc(poke, stat, statIV = 31, statEV = 0, lvl = 50, nature = "Serious")
}
\arguments{
\item{poke}{string - the identifier for the desire pokemon calculation}

\item{stat}{string - the stat for which the calculation will be performed}

\item{statIV}{integer - the individual value of the stat for the desired pokemon. Range is [0,31], default is 31}

\item{statEV}{integer - the effort value points for the desired stat. Range is [0,252], default is 0}

\item{lvl}{integer - the level of the desired pokemon. Range is [1,100], default is 50}

\item{nature}{string - the nature of the desired pokemon. Default is the stat neutral nature 'Serious'}
}
\value{
statValue
}
\description{
Calculates the stat value for a given pokemon based on their individual value for the stat, their effort value points for the stat, their level, and their nature
}
\examples{
\dontrun{
statCalc(poke,stat,statIV,statEV,lvl,nature)
}
}
\keyword{stat,calc,calculate}
