% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worlds23Data.R
\docType{data}
\name{worlds23Data}
\alias{worlds23Data}
\title{Worlds 2023 player, team, and placement data}
\format{
dataframe
}
\usage{
worlds23Data
}
\description{
Contains data on the top 113 placements from day 2 of the Pokemon Worlds Championship 2023. Includes Swiss round wins, losses, team composition, player first names, last names, and in-game names as well as country of origin.
}
\keyword{datasets}
